Re-implementation of the deletion confirmation page
This commit re-implements the deletion confirmation page due to
requests received to re-implement it.

It also adds a configuration page with a setting for
enabling|disabling the block checking whether javascript is
enabled in the user's browser prior to displaying the block itself.
This setting can be locked and customized for each user.