Correct Moodle code pre-check and lint warnings/errors
This corrects most of the pre-check lint warnings in the javascript,
and the phpdoc warnings. This involved a major re-factoring of the
javascript, which has been tested in both Chrome (current version)
and Firefox (v45.9.0). It has also been tested with the following
course formats: weeks, topics, onetopic, flexsections, and topcoll.